feat: Update Playwright configuration, reorganize OAuthProvider tests
- use `*.spec.ts` to identify e2e tests instead of `integrations` folder
